pyproject.toml
[project]
name = "agent-context"
version = "0.1.0"
description = "Search across scattered documentation: GitHub, Google Workspace, Slack, and more."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["cli", "tui", "search", "documentation", "knowledge-management", "agent-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "typer[all]>=0.12.0",
    "textual>=0.75.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.2",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "sentence-transformers>=3.0.0",
    "numpy>=2.0.0",
    "rich>=13.7.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.30.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
agent-context = "agent_context.__main__:main"

[project.entry-points."agent_context.plugins"]
github = "agent_context.plugins.github:GitHubPlugin"
google = "agent_context.plugins.google:GooglePlugin"
slack = "agent_context.plugins.slack:SlackPlugin"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_context"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
